requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
requests
httpx
numpy
scikit-learn
sentence-transformers
torch
faiss-cpu
python-multipart
jinja2
loguru
google-genai
google-generativeai
pdfplumber
